Se hizo encabezado y se separo grid
